+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 <!--next-version-placeholder-->
 
+## v2.9.1 (2025-04-30)
+
+### Fix
+
+* Bump h11 from 0.14.0 to 0.16.0 ([#132](https://github.com/ni/nisystemlink-clients-python/issues/132)) ([`2e918d6`](https://github.com/ni/nisystemlink-clients-python/commit/2e918d6b8c6daaa7ca1956bff72f0dc41d8be989))
+* Prevent string discrete values which are string of numbers from passing as numeric condition values. ([#130](https://github.com/ni/nisystemlink-clients-python/issues/130)) ([`40b68af`](https://github.com/ni/nisystemlink-clients-python/commit/40b68af4f60d779f90748e50c88366c7d6501b63))
+
+## v2.9.0 (2025-04-22)
+
+### Feature
+
+* Make TLS/SSL certificate verification optional ([#126](https://github.com/ni/nisystemlink-clients-python/issues/126)) ([`f8063f9`](https://github.com/ni/nisystemlink-clients-python/commit/f8063f9da81b57298ae0d48aee2f11e59f497b19))
+
+### Fix
+
+* Ignore extra fields in Json Model ([#131](https://github.com/ni/nisystemlink-clients-python/issues/131)) ([`4e835bb`](https://github.com/ni/nisystemlink-clients-python/commit/4e835bbbb8bd5c286acee6757d844f03957e5f05))
+
 ## v2.8.0 (2025-03-26)
 
 ### Feature